Safer/cleaner JMPENV accounting
Eval blocks have the field 'cur_top_env' which stores the value of
PL_top_env when that eval context was created. So there's no need to
do any accounting.

Plus it is not thread-safe:
  "Your direct use of the op_sv field to store the AV holding the label and
depth isn't safe on threaded builds.  On threaded builds, ops are shared
between threads, but SVs are not.  So SVs attached to ops (c.f. OP_CONST)
are moved into the pad with the op storing the pad offset (see padop
verses svop). So each interpreter gets its own pad and so its own copy of
the SV."